+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.1.4
+-----
+
+* Restructured code. Splitted the ui and runner parts.
+* Testing also against Python versions 3.2, 3.3 and 3.5.
+* Added ``name`` option to command line for running only a specific job. Even if its not enabled.
+* The config file takes a ``days`` list for a job. Before running it will check the day its running and if its in the list. Else it will skip it.
+
 0.1.3
 -----